refactor(quic): improve type clarity in SocketQUIC_error_string cast
Changed cast from unsigned int to uint8_t in SocketQUIC_error_string()
to make the 8-bit nature of QUIC_CRYPTO_ALERT explicit. The macro masks
the value to 0xff (0-255 range), so casting to uint8_t more accurately
reflects the actual value range and improves code clarity.

Closes #959

+++ src/quic/SocketQUICError.c
@@ -67,7 +67,7 @@ SocketQUIC_error_string (uint64_t code)
   if (QUIC_IS_CRYPTO_ERROR (code))
     {
       int ret = snprintf (crypto_buf, sizeof (crypto_buf), "CRYPTO_ERROR(0x%02x)",
-                          (unsigned int)QUIC_CRYPTO_ALERT (code));
+                          (uint8_t)QUIC_CRYPTO_ALERT (code));
 
       /* Defensive check: should never happen with current format */
       if (ret < 0 || (size_t)ret >= sizeof (crypto_buf))
